Gracefully handle module graph when no cabal file
Do not cause an error when asking for the module graph when there is no
cabal file.

If there is no cabal file, multi-file refactorings become impossible, as
we do not have a meaningful module graph.

We only ever use the clientModsAndFiles, which by definition means the
modules that import the one being refactored. But these will not be
loaded, so cannot be determined from the GHC module graph.